HAL: reducing .h dependency tree
# Conflicts:
#	lib/nfc_protocols/mifare_ultralight.c
