+++ applications/gui/elements.c
@@ -1,20 +1,21 @@
 #include "elements.h"
 #include "m-core.h"
 #include <assets_icons.h>
-#include "furi_hal_resources.h"
+#include <furi.h>
 #include <furi_hal.h>
-#include "gui/canvas.h"
+#include <furi_hal_resources.h>
 
-#include <gui/icon_i.h>
-#include <gui/icon_animation_i.h>
+#include "canvas.h"
+#include "icon_i.h"
+#include "icon_animation_i.h"
+#include "canvas_i.h"
 
 #include <m-string.h>
-#include <furi.h>
-#include "canvas_i.h"
 
 #include <string.h>
 #include <stdint.h>
 #include <stdbool.h>
+#include <math.h>
 
 typedef struct {
     uint8_t x;
