+++ applications/music_player/music_player_worker.c
@@ -7,6 +7,7 @@
 #include <lib/flipper_format/flipper_format.h>
 
 #include <m-array.h>
+#include <math.h>
 
 #define TAG "MusicPlayerWorker"
