+++ firmware/targets/f7/furi_hal/furi_hal_clock.c
@@ -1,11 +1,13 @@
 #include <furi_hal_clock.h>
 #include <furi.h>
 
+#include <stm32wbxx_ll_system.h>
 #include <stm32wbxx_ll_pwr.h>
 #include <stm32wbxx_ll_rcc.h>
 #include <stm32wbxx_ll_utils.h>
 #include <stm32wbxx_ll_cortex.h>
 #include <stm32wbxx_ll_bus.h>
+#include <stm32wbxx_ll_exti.h>
 
 #define TAG "FuriHalClock"
