+++ firmware/targets/f7/furi_hal/furi_hal_i2c_config.c
@@ -1,6 +1,9 @@
 #include "furi_hal_i2c_config.h"
+#include "furi_hal_i2c_i.h"
 #include <furi_hal_resources.h>
 #include <furi_hal_version.h>
+
+#include <stm32wbxx_ll_i2c.h>
 #include <stm32wbxx_ll_bus.h>
 #include <stm32wbxx_ll_rcc.h>
