+++ firmware/targets/f7/furi_hal/furi_hal_i2c_i.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <stm32wbxx_ll_i2c.h>
+
+/** FuriHal i2c bus */
+struct FuriHalI2cBus {
+    I2C_TypeDef* i2c;
+    FuriHalI2cBusHandle* current_handle;
+    FuriHalI2cBusEventCallback callback;
+};
+
