+++ firmware/targets/f7/furi_hal/furi_hal_idle_timer_i.h
