+++ firmware/targets/f7/furi_hal/furi_hal_resources.c
@@ -4,6 +4,7 @@
 
 #include <stm32wbxx_ll_rcc.h>
 #include <stm32wbxx_ll_pwr.h>
+#include <stm32wbxx_ll_gpio.h>
 
 const GpioPin vibro_gpio = {.port = VIBRO_GPIO_Port, .pin = VIBRO_Pin};
 const GpioPin ibutton_gpio = {.port = iBTN_GPIO_Port, .pin = iBTN_Pin};
