+++ firmware/targets/f7/furi_hal/furi_hal_spi_config.h
@@ -2,6 +2,8 @@
 
 #include <furi_hal_spi_types.h>
 
+#include <stm32wbxx_ll_spi.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
