+++ firmware/targets/f7/furi_hal/furi_hal_spi_types.h
@@ -5,10 +5,6 @@
 
 #include <furi_hal_gpio.h>
 
-#include <stm32wbxx_ll_spi.h>
-#include <stm32wbxx_ll_rcc.h>
-#include <stm32wbxx_ll_bus.h>
-
 #ifdef __cplusplus
 extern "C" {
 #endif
