+++ firmware/targets/furi_hal_include/furi_hal_usb.h
@@ -1,23 +1,6 @@
 #pragma once
 
-#include "usb.h"
-
 typedef struct FuriHalUsbInterface FuriHalUsbInterface;
-
-struct FuriHalUsbInterface {
-    void (*init)(usbd_device* dev, FuriHalUsbInterface* intf, void* ctx);
-    void (*deinit)(usbd_device* dev);
-    void (*wakeup)(usbd_device* dev);
-    void (*suspend)(usbd_device* dev);
-
-    struct usb_device_descriptor* dev_descr;
-
-    void* str_manuf_descr;
-    void* str_prod_descr;
-    void* str_serial_descr;
-
-    void* cfg_descr;
-};
 
 /** USB device interface modes */
 extern FuriHalUsbInterface usb_cdc_single;
