+++ lib/nfc_protocols/mifare_ultralight.c
@@ -2,6 +2,7 @@
 #include "mifare_ultralight.h"
 #include <furi.h>
 #include <m-string.h>
+#include <nfc_common.h>
 
 #define TAG "MfUltralight"
 
@@ -1103,7 +1104,7 @@ bool mf_ul_prepare_emulation_response(
                 if(buff_rx_len == 1 * 8) {
                     tx_bytes = sizeof(emulator->data.version);
                     memcpy(buff_tx, &emulator->data.version, tx_bytes);
-                    *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                    *data_type = NFC_TXRX_DEFAULT;
                     command_parsed = true;
                 }
             }
@@ -1207,7 +1208,7 @@ bool mf_ul_prepare_emulation_response(
                             if(ascii_mirror_cptr != NULL) {
                                 string_clear(ascii_mirror);
                             }
-                            *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                            *data_type = NFC_TXRX_DEFAULT;
                             command_parsed = true;
                         } while(false);
                     }
@@ -1243,7 +1244,7 @@ bool mf_ul_prepare_emulation_response(
                                     &buff_tx[12], &emulator->data.data[(start_page + 1) * 4], 4);
                             mf_ul_protect_auth_data_on_read_command_i2c(
                                 buff_tx, start_page, start_page + copy_count / 4 - 1, emulator);
-                            *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                            *data_type = NFC_TXRX_DEFAULT;
                             command_parsed = true;
                         }
                     } else {
@@ -1260,7 +1261,7 @@ bool mf_ul_prepare_emulation_response(
                             // the read out, while if you read the mirrored session registers,
                             // it returns both session registers on either pages
                             memset(buff_tx, 0, tx_bytes);
-                            *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                            *data_type = NFC_TXRX_DEFAULT;
                             command_parsed = true;
                             emulator->ntag_i2c_plus_sector3_lockout = true;
                         }
@@ -1352,7 +1353,7 @@ bool mf_ul_prepare_emulation_response(
                                                 memset(&buff_tx[(pwd_page_offset + 1) * 4], 0, 4);
                                         }
                                     }
-                                    *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                                    *data_type = NFC_TXRX_DEFAULT;
                                     command_parsed = true;
                                 } while(false);
                             }
@@ -1376,7 +1377,7 @@ bool mf_ul_prepare_emulation_response(
                                         start_page,
                                         start_page + copy_count / 4 - 1,
                                         emulator);
-                                    *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                                    *data_type = NFC_TXRX_DEFAULT;
                                     command_parsed = true;
                                 }
                             }
@@ -1464,7 +1465,7 @@ bool mf_ul_prepare_emulation_response(
                             buff_tx[1] = (emulator->data.counter[cnt_num] >> 8) & 0xFF;
                             buff_tx[2] = (emulator->data.counter[cnt_num] >> 16) & 0xFF;
                             tx_bytes = 3;
-                            *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                            *data_type = NFC_TXRX_DEFAULT;
                             command_parsed = true;
                         }
                     } while(false);
@@ -1500,7 +1501,7 @@ bool mf_ul_prepare_emulation_response(
                         // AUTHLIM reached, always fail
                         buff_tx[0] = MF_UL_NAK_AUTHLIM_REACHED;
                         tx_bits = 4;
-                        *data_type = FURI_HAL_NFC_TX_RAW_RX_DEFAULT;
+                        *data_type = NFC_TX_RAW_RX_DEFAULT;
                         mf_ul_reset_emulation(emulator, false);
                         command_parsed = true;
                     } else {
@@ -1509,7 +1510,7 @@ bool mf_ul_prepare_emulation_response(
                             buff_tx[0] = emulator->config->auth_data.pack.raw[0];
                             buff_tx[1] = emulator->config->auth_data.pack.raw[1];
                             tx_bytes = 2;
-                            *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                            *data_type = NFC_TXRX_DEFAULT;
                             emulator->auth_success = true;
                             command_parsed = true;
                             if(emulator->data.curr_authlim != 0) {
@@ -1522,7 +1523,7 @@ bool mf_ul_prepare_emulation_response(
                             buff_tx[0] = 0x80;
                             buff_tx[1] = 0x80;
                             tx_bytes = 2;
-                            *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                            *data_type = NFC_TXRX_DEFAULT;
                             emulator->auth_success = true;
                             command_parsed = true;
                         } else {
@@ -1536,7 +1537,7 @@ bool mf_ul_prepare_emulation_response(
                                 emulator->data.curr_authlim = UINT16_MAX;
                                 buff_tx[0] = MF_UL_NAK_AUTHLIM_REACHED;
                                 tx_bits = 4;
-                                *data_type = FURI_HAL_NFC_TX_RAW_RX_DEFAULT;
+                                *data_type = NFC_TX_RAW_RX_DEFAULT;
                                 mf_ul_reset_emulation(emulator, false);
                                 command_parsed = true;
                             } else {
@@ -1552,7 +1553,7 @@ bool mf_ul_prepare_emulation_response(
                 if(buff_rx_len == (1 + 1) * 8 && buff_rx[1] == 0x00) {
                     tx_bytes = sizeof(emulator->data.signature);
                     memcpy(buff_tx, emulator->data.signature, tx_bytes);
-                    *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                    *data_type = NFC_TXRX_DEFAULT;
                     command_parsed = true;
                 }
             }
@@ -1563,7 +1564,7 @@ bool mf_ul_prepare_emulation_response(
                     if(cnt_num < 3) {
                         buff_tx[0] = emulator->data.tearing[cnt_num];
                         tx_bytes = 1;
-                        *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                        *data_type = NFC_TXRX_DEFAULT;
                         command_parsed = true;
                     }
                 }
@@ -1585,7 +1586,7 @@ bool mf_ul_prepare_emulation_response(
                 if(buff_rx_len == (1 + 20) * 8) {
                     buff_tx[0] = emulator->config_cache.vctid;
                     tx_bytes = 1;
-                    *data_type = FURI_HAL_NFC_TXRX_DEFAULT;
+                    *data_type = NFC_TXRX_DEFAULT;
                     command_parsed = true;
                 }
             }
@@ -1608,18 +1609,18 @@ bool mf_ul_prepare_emulation_response(
         // Send NACK
         buff_tx[0] = MF_UL_NAK_INVALID_ARGUMENT;
         tx_bits = 4;
-        *data_type = FURI_HAL_NFC_TX_RAW_RX_DEFAULT;
+        *data_type = NFC_TX_RAW_RX_DEFAULT;
         // Every NAK should cause reset to IDLE
         mf_ul_reset_emulation(emulator, false);
     } else if(send_ack) {
         buff_tx[0] = MF_UL_ACK;
         tx_bits = 4;
-        *data_type = FURI_HAL_NFC_TX_RAW_RX_DEFAULT;
+        *data_type = NFC_TX_RAW_RX_DEFAULT;
     }
 
     if(respond_nothing) {
         *buff_tx_len = UINT16_MAX;
-        *data_type = FURI_HAL_NFC_TX_RAW_RX_DEFAULT;
+        *data_type = NFC_TX_RAW_RX_DEFAULT;
     } else {
         // Return tx buffer size in bits
         if(tx_bytes) {
